requirements.txt
pandas>=1.3.0
python-dotenv>=0.19.0
requests>=2.26.0
tweepy>=4.10.0
numpy>=1.20.0